+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - `COMPILER_INDEX_STORE_ENABLE` is now forcefully enabled as it's required for indexing.
 - Fix false positive where a `typealias` is extended but otherwise unused.
 - Fix redundant accessibility analysis for function metatype arguments.
+- Fix redundant accessibility analysis for property types inferred from a function call initializer.
 
 ## 2.12.1 (2023-03-04)